Secured the api with authentication and authorization

+++ route/api/genres.js
@@ -1,7 +1,10 @@
 const mongoose = require('mongoose');
 const express = require('express');
 const Joi = require('joi');
+const auth = require('../../middleware/auth');
+const admin = require('../../middleware/admin');
 const { Genre, validateGenre } = require('../models/genres');
+const { validateId } = require('../util/validator');
 const router = express.Router();
   
   router.get('/', async (req, res) => {
@@ -9,7 +12,7 @@ const router = express.Router();
     res.send(genres);
   });
   
-  router.post('/', async (req, res) => {
+  router.post('/', auth, async (req, res) => {
     const { error } = validateGenre(req.body); 
     if (error) return res.status(400).send(error.details[0].message);
 
@@ -19,8 +22,9 @@ const router = express.Router();
     res.send(genre);
   });
   
-  router.put('/:id', async (req, res) => {
+  router.put('/:id', auth ,async (req, res) => {
     if (!req.params.id) return res.status(400).send('Please send genre id as req param');
+    if (!validateId(req.params.id)) return res.status(400).send('Please send valid genre id as req param');
     const { error } = validateGenre(req.body); 
     if (error) return res.status(400).send(error.details[0].message);
 
@@ -32,8 +36,9 @@ const router = express.Router();
     res.send(genre);
   });
   
-  router.delete('/:id', async (req, res) => {
+  router.delete('/:id', [auth, admin], async (req, res) => {
     if (!req.params.id) return res.status(400).send('Please send genre id as req param');
+    if (!validateId(req.params.id)) return res.status(400).send('Please send valid genre id as req param');
     const genre = await Genre.findByIdAndDelete(req.params.id);
     if (!genre) return res.status(404).send('The genre with the given ID was not found.');
   
@@ -42,6 +47,7 @@ const router = express.Router();
   
   router.get('/:id', async (req, res) => {
     if (!req.params.id) return res.status(400).send('Please send genre id as req param');
+    if (!validateId(req.params.id)) return res.status(400).send('Please send valid genre id as req param');
     const genre = await Genre.findById(req.params.id);
     if (!genre) return res.status(404).send('The genre with the given ID was not found.');
     res.send(genre);
